+++ tawazi/node.py
@@ -1,12 +1,12 @@
-import inspect
 from copy import deepcopy
 from threading import Lock
 from types import MethodType
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Union
 
 from loguru import logger
 
 from tawazi.errors import TawaziBaseException, UnvalidExecNodeCall, raise_arg_exc
+from tawazi.helpers import ordinal
 
 from .config import Cfg
 
@@ -25,63 +25,6 @@ ARG_NAME_SEP = ">>>"
 
 class NoVal:
     ...
-
-
-# TODO: make a helpers module and put this function in it
-def ordinal(numb: int) -> str:
-    if numb < 20:  # determining suffix for < 20
-        if numb == 1:
-            suffix = "st"
-        elif numb == 2:
-            suffix = "nd"
-        elif numb == 3:
-            suffix = "rd"
-        else:
-            suffix = "th"
-    else:  # determining suffix for > 20
-        tens = str(numb)
-        tens = tens[-2]
-        unit = str(numb)
-        unit = unit[-1]
-        if tens == "1":
-            suffix = "th"
-        else:
-            if unit == "1":
-                suffix = "st"
-            elif unit == "2":
-                suffix = "nd"
-            elif unit == "3":
-                suffix = "rd"
-            else:
-                suffix = "th"
-    return str(numb) + suffix
-
-
-# TODO: transfer to a helpers file
-# TODO: change the name!!
-def get_args_and_default_args(func: Callable[..., Any]) -> Tuple[List[str], Dict[str, Any]]:
-    """
-    Retrieves the arguments names and the default arguments of a function.
-    Args:
-        func: the target function
-
-    Returns:
-        A Tuple containing a List of argument names of non default arguments,
-         and the mapping between the arguments and their default value for default arguments
-    >>> def f(a1, a2, *args, d1=123, d2=None): pass
-    >>> get_args_and_default_args(f)
-    >>> (['a1', 'a2', 'args'], {'d1': 123, 'd2': None})
-    """
-    signature = inspect.signature(func)
-    args = []
-    default_args = {}
-    for k, v in signature.parameters.items():
-        if v.default is not inspect.Parameter.empty:
-            default_args[k] = v.default
-        else:
-            args.append(k)
-
-    return args, default_args
 
 
 class ExecNode:
@@ -276,8 +219,6 @@ class LazyExecNode(ExecNode):
             setup=setup,
         )
 
-    # TODO: support tagging an ExecNode using __tawazi_call_tag or smthing like that
-    #  maybe something like twz_tag might be acceptable ?
     def __call__(self, *args: Any, **kwargs: Any) -> "LazyExecNode":
         """
         Record the dependencies in a global variable to be called later in DAG.
