+++ tawazi/ops.py
@@ -3,17 +3,11 @@ from typing import Any, Callable, List, Optional, Union
 
 from tawazi import DAG
 from tawazi.errors import ErrorStrategy
+from tawazi.helpers import get_args_and_default_args
 
 from . import node
 from .config import Cfg
-from .node import (
-    ArgExecNode,
-    ExecNode,
-    IdentityHash,
-    LazyExecNode,
-    exec_nodes_lock,
-    get_args_and_default_args,
-)
+from .node import ArgExecNode, ExecNode, IdentityHash, LazyExecNode, exec_nodes_lock
 
 
 # TODO: modify is_sequential's default value according to the pre used default
